Completed Pass-Through functionality on MiddleWare
The Middleware now successfully acts as a pass-through from the
front-end (iOS application) to the backend (php and MySQL).
This commit is the current state of the Middleware located on
the Atlantic.net server (domain soon to point there).

Next steps will be to create the logic on top of the
data flowing through the middleware in order to
do interesting things with it.

Also a gitignore was added. So much gitignore was ignored in previous
commits.

Left a few files untracked...I don't know what they are, XCode related
but I shouldn't have tried messing with the xcode stuff.

Current time: Midnight